LCA Java first commit
After the hurricane power outages, (and leaving my external laptop charger in college during the hurricane) I decide that to do this project in Haskell, a language that I'm unfamiliar with, would be infeasible. I just don't have the time to learn of to use HUnit or Test Framework in time. Agile.

Will update with test framework asap.